requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.37.0
requests==2.32.3
